DWYN/cogs/utility/ping.py
# Primitive as hell ping. Will change to a genuinely functioning system at some point.
from discord.ext import commands

class Ping(commands.Cog):
    def __init__(self, bot):
        self.bot = bot

    @commands.command()
    async def ping(self, ctx):
        await ctx.send("Pong!")

async def setup(bot):
    await bot.add_cog(Ping(bot))
