requirements.txt
discord-py
python-dotenv
audioop-lts
asyncio